Add complex number support (TYPE_COMPLEX) and fix FSet quickload crash
The FSet quickload crash at code_pos=29227 was caused by #C(5 16) complex
number reader syntax producing a dotted pair (5 . 16) instead of a proper
complex number type. The compiler correctly rejected this as a malformed
call form. Previously masked because Test-Misc-2 was skipped due to the
64KB bytecode limit.

Changes:
- Add TYPE_COMPLEX heap type with CL_Complex struct (realpart, imagpart)
- Reader: #C(real imag) now produces TYPE_COMPLEX objects
- Printer: complex numbers print as #C(real imag)
- Compiler: complex numbers are self-evaluating constants
- Builtins: complex, complexp, realpart, imagpart, conjugate
- EQL/EQUAL: component-wise comparison for complex numbers
- Hash tables: hash and equality support for complex keys
- FASL: serialization/deserialization (FASL_TAG_COMPLEX)
- Type system: typep, type-of, numberp, CL_NUMBER_P updated
- GC: mark and verify for complex number components
- Fix gc_root_count leak in bi_load, compile-file, and FASL load error
  paths (CL_CATCH longjmp skips CL_GC_UNPROTECT, leaking stale roots)
- Add GC protection to compiler iteration patterns (compile_setf,
  funcall, apply, case, typecase, flet, labels, handler-bind,
  restart-case, macrolet, multiple-value-prog1)
- i32 jump offsets: 64KB→256KB bytecode limit per function
- 27 new complex number host tests + 19 Amiga tests

+++ src/core/vm.h
@@ -67,7 +67,7 @@ typedef struct {
     CL_Obj tag;            /* catch tag (CATCH only) */
     CL_Obj result;         /* value to propagate */
     uint32_t catch_ip;     /* IP after the CATCH/UWPROT instruction */
-    int16_t offset;        /* jump offset to landing */
+    int32_t offset;        /* jump offset to landing */
     uint8_t *code;         /* code pointer to restore */
     CL_Obj *constants;     /* constants pointer to restore */
     CL_Obj bytecode;       /* frame->bytecode at NLX push time */
